fix(cache): invalidate categories cache when articles are imported or created

+++ src/NewsPortal.Service/Services/CategoryService.cs
@@ -42,7 +42,7 @@ public class CategoryService : ICategoryService
                 Color = c.Color,
                 ArticleCount = c.Articles.Count
             }).ToList();
-        }, CacheDurations.Long);
+        }, CacheDurations.Short);
     }
 
     public async Task<CategoryDto?> GetCategoryBySlugAsync(string slug)
